+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - feat: add python 3.10 support (#232)
+- Contributing, contributors & code of conduct files (#236)
 
 ## [0.36.0] - 2022-11-22