+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update test after test only for datasample removal (#235)
 
+### Added
+
+- Remote test to verify that, with the right permission, we can test on an org from the training model of an other organization (#237)
+
 ## [0.37.0] - 2022-12-19
 
 ### Added