+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remote test to verify that, with the right permission, we can test on an org from the training model of an other organization (#237)
 
+### Changed
+
+- chore: rename tuple to task (#239)
+
 ## [0.38.0] - 2023-01-30
 
 ### Changed