+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## [0.38.1] - 2023-06-02
+## [0.38.1] - 2023-02-06
 
 ### Added